+++ jit/src/lib.rs
@@ -1,0 +1,5 @@
+mod lexer;
+mod parser;
+mod ir;
+mod codegen;
+// mod vm;
